+++ src/test/java/com/nubiform/payment/api/InvalidParamTest.java
@@ -18,7 +18,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 @SpringBootTest
 @AutoConfigureMockMvc
-class PaymentControllerInvalidParamTest {
+class InvalidParamTest {
 
     @Autowired
     MockMvc mockMvc;
