--- src/test/java/com/nubiform/payment/api/PaymentControllerCancelTest.java
@@ -1,86 +1,0 @@
-package com.nubiform.payment.api;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.nubiform.payment.repository.BalanceRepository;
-import com.nubiform.payment.repository.SentRepository;
-import com.nubiform.payment.service.PaymentService;
-import com.nubiform.payment.vo.CancelRequest;
-import com.nubiform.payment.vo.SubmitRequest;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.http.MediaType;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.MvcResult;
-
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
-import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
-@SpringBootTest
-@AutoConfigureMockMvc
-class PaymentControllerCancelTest {
-
-    @Autowired
-    MockMvc mockMvc;
-
-    @Autowired
-    PaymentService paymentService;
-
-    @Autowired
-    BalanceRepository balanceRepository;
-
-    @Autowired
-    SentRepository sentRepository;
-
-    @Autowired
-    ObjectMapper objectMapper;
-
-    SubmitRequest submitRequest;
-    CancelRequest cancelRequest;
-
-    @BeforeEach
-    void setUp() throws Exception {
-        submitRequest = new SubmitRequest();
-        submitRequest.setCard("1234567890123456");
-        submitRequest.setExpiration("1234");
-        submitRequest.setCvc("123");
-        submitRequest.setInstallment(0);
-        submitRequest.setAmount(10000L);
-
-        Long id = paymentService.submit(submitRequest).getId();
-
-        cancelRequest = new CancelRequest();
-        cancelRequest.setId(id);
-        cancelRequest.setAmount(10000L);
-    }
-
-    @Test
-    public void delPayment() throws Exception {
-        MvcResult mvcResult = mockMvc.perform(delete("/api/v1/payment")
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(objectMapper.writeValueAsString(cancelRequest)))
-                .andDo(print())
-                .andExpect(status().isOk())
-                .andReturn();
-
-        String responseBody = mvcResult.getResponse().getContentAsString();
-
-        System.out.println(balanceRepository.findById(1L).orElse(null));
-    }
-
-    @Test
-    public void delPaymentFailure() throws Exception {
-        cancelRequest.setAmount(1000000L);
-
-        mockMvc.perform(delete("/api/v1/payment")
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(objectMapper.writeValueAsString(cancelRequest)))
-                .andDo(print())
-                .andExpect(jsonPath("$.code").exists())
-                .andExpect(jsonPath("$.message").exists());
-    }
-}
