+++ src/test/java/com/nubiform/payment/api/SubmitConcurrencyTest.java
@@ -1,9 +1,7 @@
 package com.nubiform.payment.api;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.nubiform.payment.domain.Sent;
 import com.nubiform.payment.repository.SentRepository;
-import com.nubiform.payment.vo.Response;
 import com.nubiform.payment.vo.SubmitRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -11,18 +9,23 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.MediaType;
+import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.MvcResult;
 
-import static org.junit.jupiter.api.Assertions.assertNotNull;
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
+@ActiveProfiles("test")
 @SpringBootTest
 @AutoConfigureMockMvc
-class PaymentControllerSubmitTest {
+class SubmitConcurrencyTest {
+
+    public static final int N_THREADS = 100;
 
     @Autowired
     MockMvc mockMvc;
@@ -47,19 +50,23 @@ class PaymentControllerSubmitTest {
 
     @Test
     public void postPayment() throws Exception {
-        MvcResult mvcResult = mockMvc.perform(post("/api/v1/payment")
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(objectMapper.writeValueAsString(submitRequest)))
-                .andDo(print())
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.id").exists())
-                .andExpect(jsonPath("$.data").exists())
-                .andReturn();
+        ExecutorService executorService = Executors.newFixedThreadPool(N_THREADS);
+        CountDownLatch countDownLatch = new CountDownLatch(N_THREADS);
 
-        String responseBody = mvcResult.getResponse().getContentAsString();
-        Response response = objectMapper.readValue(responseBody, Response.class);
-        Sent sent = sentRepository.findById(response.getLongId()).orElse(null);
-
-        assertNotNull(sent);
+        for (int i = 0; i < N_THREADS; i++) {
+            executorService.execute(() -> {
+                try {
+                    MvcResult mvcResult = mockMvc.perform(post("/api/v1/payment")
+                            .contentType(MediaType.APPLICATION_JSON)
+                            .content(objectMapper.writeValueAsString(submitRequest)))
+                            .andDo(print())
+                            .andReturn();
+                } catch (Exception e) {
+                    e.printStackTrace();
+                }
+                countDownLatch.countDown();
+            });
+        }
+        countDownLatch.await();
     }
 }
